test_task_ivi/requirements.txt
pytest==6.2.2
requests==2.26.0
selenium==4.0.0
python-dotenv==0.19.2
pytest-html==3.1.1